seperate typespec project check into own class
